+++ lib/swatches.dart
@@ -3,5 +3,6 @@
 export 'src/material/blue_greys.dart';
 export 'src/material/browns.dart';
 export 'src/material/greys.dart';
+export 'src/material/primaries.dart';
 export 'src/swatch.dart';
 export 'src/swatch_base.dart';