+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added [#24](https://github.com/dartoos-dev/eo_color/issues/24)
+- Partial implementation of the Primaries swatch.
+
 ### Changed [#25](https://github.com/dartoos-dev/eo_color/issues/25)
 - Package logo in README - cute wooden color palette.